requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
tqdm==4.66.1
Pillow==10.1.0
seaborn==0.13.0
jupyterlab==4.0.9
